add/rename: add SessionResult type and rename VM
