+++ KikurageFeature/QRCodeReader/SessionSetupResult.swift
@@ -1,0 +1,9 @@
+//
+//  SessionSetupResult.swift
+//  KikurageFeature
+//
+//  Created by Shusuke Ota on 2022/7/24.
+//  Copyright © 2022 shusuke. All rights reserved.
+//
+
+import Foundation
